POT Threshold Options IN PROGRESS
Added options for POT thresholds - Automatic, Choose, Manual. At the moment Automatic works for non-RFA POT, but I still need to figure out how to generate the choose list, and there seems to be errors for entering in a manual number (despite using 20, which is was before..)
It seems RFA-POT doesn't have any sort of parameter for threshold, and that it's automatically taking the thresh from the dataset given. Is there a way I should be inputting this? Or should we rewrite how FloodnetPool handles threshold?