Dynamic lists for Supreg and Thresholds
for Rfa-type models, supreg and threshold options will now populate from the Station Data file that is read in. Clickupdate configured to allow any supreg and threshold now, not just hardcoded for the types we were working on.
Minor fixes to how distribution/threshold values are handled in the two functions.