SpacePlots + Some Import/Documentation
Updated the spacePlots to more-properly show what they are supposed to show. They are now generated when fitting is done, and are saved with the models, so future loading/swapping between displayed functions will be much smoother now.

Tried getting the imports to work. I think that since app.R isn't included in the R folder, roxygen isn't picking up the import documentation, so I added it to the floodnetApp() function which calls the shiny app. It seems like I'll need to keep the library() calls in app.R, but this should be okay since the functions are now imported in the namespace due to their inclusion in floodnetApp.R